+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: (PLS) Integrate new v2 API for Pulsechain as a pulsechain-scan adapter
+- added: (Tron) Add energy penalties to fee calculation
 - fixed: (PLS) update evmscan adapter servers for Pulsechain to fix transaction history queries
 
 ## 3.1.2 (2024-02-20)