+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.1.2-1 (2024-02-23)
+
 - added: (Tron) Add energy penalties to fee calculation
 - added: Add more Arbitrum tokens